// src/dto/create.customer.dto.ts
export interface CreateCustomerDto {
  name: string;
  subscription: string;
  email: string;
  phone: string;
  company: string;
  postalCode: string;
  address: string;
  addressNumber: string;
  addressDistrict: string;
  city: string;
  country: string;
  birthDate: string;
  noteAbout: string;
  userId: string;
  // typeSubscription: number;
}